feat(profile): Enhance profile picture handling in ApplicationAnswerFile model and views
- Added logic to save and delete profile pictures correctly when associated with application answer files.
- Updated the delete view to prevent file deletion if the application is already completed.
- Improved file download handling to support profile pictures stored in the CustomUser model.
- Enhanced the template to display download buttons for file answers more effectively.

+++ FWMsg/BW/views.py
@@ -231,6 +231,11 @@ def bw_application_file_answer(request, file_question_id):
 @required_role('BO')
 @application_or_seminar_is_open
 def bw_application_file_answer_delete(request, file_answer_id):
+    bewerber = Bewerber.objects.get(user=request.user)
+    if bewerber.abgeschlossen == True:
+        messages.error(request, 'Du hast bereits Deine Bewerbung abgeschlossen und kannst keine Dateien mehr löschen.')
+        return redirect('bw_home')
+
     file_answer = ApplicationAnswerFile.objects.get(id=file_answer_id, user=request.user)
     file_answer.delete()
     messages.success(request, 'Datei wurde erfolgreich gelöscht')
